+++ BUILD
@@ -2189,7 +2189,6 @@ grpc_cc_library(
 grpc_cc_library(
     name = "iomgr_timer",
     srcs = [
-        "src/core/lib/iomgr/time_averaged_stats.cc",
         "src/core/lib/iomgr/timer.cc",
         "src/core/lib/iomgr/timer_generic.cc",
         "src/core/lib/iomgr/timer_heap.cc",
@@ -2200,7 +2199,6 @@ grpc_cc_library(
         "src/core/lib/iomgr/timer_generic.h",
         "src/core/lib/iomgr/timer_heap.h",
         "src/core/lib/iomgr/timer_manager.h",
-        "src/core/lib/iomgr/time_averaged_stats.h",
     ] + [
         # TODO(hork): deduplicate
         "src/core/lib/iomgr/iomgr.h",
@@ -2218,6 +2216,7 @@ grpc_cc_library(
         "grpc_trace",
         "iomgr_port",
         "time",
+        "time_averaged_stats",
         "useful",
     ],
 )
@@ -2286,10 +2285,10 @@ grpc_cc_library(
 )
 
 grpc_cc_library(
-    name = "iomgr_ee_time_averaged_stats",
-    srcs = ["src/core/lib/event_engine/iomgr_engine/time_averaged_stats.cc"],
+    name = "time_averaged_stats",
+    srcs = ["src/core/lib/gprpp/time_averaged_stats.cc"],
     hdrs = [
-        "src/core/lib/event_engine/iomgr_engine/time_averaged_stats.h",
+        "src/core/lib/gprpp/time_averaged_stats.h",
     ],
     tags = ["grpc-autodeps"],
     deps = ["gpr_base"],
@@ -2313,8 +2312,8 @@ grpc_cc_library(
     deps = [
         "event_engine_base_hdrs",
         "gpr_base",
-        "iomgr_ee_time_averaged_stats",
         "time",
+        "time_averaged_stats",
         "useful",
     ],
 )
