+++ config.w32
@@ -435,7 +435,6 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\event_engine\\event_engine.cc " +
     "src\\core\\lib\\event_engine\\iomgr_engine\\iomgr_engine.cc " +
     "src\\core\\lib\\event_engine\\iomgr_engine\\thread_pool.cc " +
-    "src\\core\\lib\\event_engine\\iomgr_engine\\time_averaged_stats.cc " +
     "src\\core\\lib\\event_engine\\iomgr_engine\\timer.cc " +
     "src\\core\\lib\\event_engine\\iomgr_engine\\timer_heap.cc " +
     "src\\core\\lib\\event_engine\\iomgr_engine\\timer_manager.cc " +
@@ -486,6 +485,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\gprpp\\thd_posix.cc " +
     "src\\core\\lib\\gprpp\\thd_windows.cc " +
     "src\\core\\lib\\gprpp\\time.cc " +
+    "src\\core\\lib\\gprpp\\time_averaged_stats.cc " +
     "src\\core\\lib\\gprpp\\time_util.cc " +
     "src\\core\\lib\\gprpp\\work_serializer.cc " +
     "src\\core\\lib\\http\\format_request.cc " +
@@ -554,7 +554,6 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\iomgr\\tcp_server_utils_posix_noifaddrs.cc " +
     "src\\core\\lib\\iomgr\\tcp_server_windows.cc " +
     "src\\core\\lib\\iomgr\\tcp_windows.cc " +
-    "src\\core\\lib\\iomgr\\time_averaged_stats.cc " +
     "src\\core\\lib\\iomgr\\timer.cc " +
     "src\\core\\lib\\iomgr\\timer_generic.cc " +
     "src\\core\\lib\\iomgr\\timer_heap.cc " +
