+++ gRPC-C++.podspec
@@ -683,7 +683,6 @@ Pod::Spec.new do |s|
                       'src/core/lib/event_engine/handle_containers.h',
                       'src/core/lib/event_engine/iomgr_engine/iomgr_engine.h',
                       'src/core/lib/event_engine/iomgr_engine/thread_pool.h',
-                      'src/core/lib/event_engine/iomgr_engine/time_averaged_stats.h',
                       'src/core/lib/event_engine/iomgr_engine/timer.h',
                       'src/core/lib/event_engine/iomgr_engine/timer_heap.h',
                       'src/core/lib/event_engine/iomgr_engine/timer_manager.h',
@@ -728,6 +727,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/gprpp/table.h',
                       'src/core/lib/gprpp/thd.h',
                       'src/core/lib/gprpp/time.h',
+                      'src/core/lib/gprpp/time_averaged_stats.h',
                       'src/core/lib/gprpp/time_util.h',
                       'src/core/lib/gprpp/unique_type_name.h',
                       'src/core/lib/gprpp/work_serializer.h',
@@ -790,7 +790,6 @@ Pod::Spec.new do |s|
                       'src/core/lib/iomgr/tcp_server.h',
                       'src/core/lib/iomgr/tcp_server_utils_posix.h',
                       'src/core/lib/iomgr/tcp_windows.h',
-                      'src/core/lib/iomgr/time_averaged_stats.h',
                       'src/core/lib/iomgr/timer.h',
                       'src/core/lib/iomgr/timer_generic.h',
                       'src/core/lib/iomgr/timer_heap.h',
@@ -1522,7 +1521,6 @@ Pod::Spec.new do |s|
                               'src/core/lib/event_engine/handle_containers.h',
                               'src/core/lib/event_engine/iomgr_engine/iomgr_engine.h',
                               'src/core/lib/event_engine/iomgr_engine/thread_pool.h',
-                              'src/core/lib/event_engine/iomgr_engine/time_averaged_stats.h',
                               'src/core/lib/event_engine/iomgr_engine/timer.h',
                               'src/core/lib/event_engine/iomgr_engine/timer_heap.h',
                               'src/core/lib/event_engine/iomgr_engine/timer_manager.h',
@@ -1567,6 +1565,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/gprpp/table.h',
                               'src/core/lib/gprpp/thd.h',
                               'src/core/lib/gprpp/time.h',
+                              'src/core/lib/gprpp/time_averaged_stats.h',
                               'src/core/lib/gprpp/time_util.h',
                               'src/core/lib/gprpp/unique_type_name.h',
                               'src/core/lib/gprpp/work_serializer.h',
@@ -1629,7 +1628,6 @@ Pod::Spec.new do |s|
                               'src/core/lib/iomgr/tcp_server.h',
                               'src/core/lib/iomgr/tcp_server_utils_posix.h',
                               'src/core/lib/iomgr/tcp_windows.h',
-                              'src/core/lib/iomgr/time_averaged_stats.h',
                               'src/core/lib/iomgr/timer.h',
                               'src/core/lib/iomgr/timer_generic.h',
                               'src/core/lib/iomgr/timer_heap.h',
