+++ gRPC-Core.podspec
@@ -1050,8 +1050,6 @@ Pod::Spec.new do |s|
                       'src/core/lib/event_engine/iomgr_engine/iomgr_engine.h',
                       'src/core/lib/event_engine/iomgr_engine/thread_pool.cc',
                       'src/core/lib/event_engine/iomgr_engine/thread_pool.h',
-                      'src/core/lib/event_engine/iomgr_engine/time_averaged_stats.cc',
-                      'src/core/lib/event_engine/iomgr_engine/time_averaged_stats.h',
                       'src/core/lib/event_engine/iomgr_engine/timer.cc',
                       'src/core/lib/event_engine/iomgr_engine/timer.h',
                       'src/core/lib/event_engine/iomgr_engine/timer_heap.cc',
@@ -1146,6 +1144,8 @@ Pod::Spec.new do |s|
                       'src/core/lib/gprpp/thd_windows.cc',
                       'src/core/lib/gprpp/time.cc',
                       'src/core/lib/gprpp/time.h',
+                      'src/core/lib/gprpp/time_averaged_stats.cc',
+                      'src/core/lib/gprpp/time_averaged_stats.h',
                       'src/core/lib/gprpp/time_util.cc',
                       'src/core/lib/gprpp/time_util.h',
                       'src/core/lib/gprpp/unique_type_name.h',
@@ -1276,8 +1276,6 @@ Pod::Spec.new do |s|
                       'src/core/lib/iomgr/tcp_server_windows.cc',
                       'src/core/lib/iomgr/tcp_windows.cc',
                       'src/core/lib/iomgr/tcp_windows.h',
-                      'src/core/lib/iomgr/time_averaged_stats.cc',
-                      'src/core/lib/iomgr/time_averaged_stats.h',
                       'src/core/lib/iomgr/timer.cc',
                       'src/core/lib/iomgr/timer.h',
                       'src/core/lib/iomgr/timer_generic.cc',
@@ -2132,7 +2130,6 @@ Pod::Spec.new do |s|
                               'src/core/lib/event_engine/handle_containers.h',
                               'src/core/lib/event_engine/iomgr_engine/iomgr_engine.h',
                               'src/core/lib/event_engine/iomgr_engine/thread_pool.h',
-                              'src/core/lib/event_engine/iomgr_engine/time_averaged_stats.h',
                               'src/core/lib/event_engine/iomgr_engine/timer.h',
                               'src/core/lib/event_engine/iomgr_engine/timer_heap.h',
                               'src/core/lib/event_engine/iomgr_engine/timer_manager.h',
@@ -2177,6 +2174,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/gprpp/table.h',
                               'src/core/lib/gprpp/thd.h',
                               'src/core/lib/gprpp/time.h',
+                              'src/core/lib/gprpp/time_averaged_stats.h',
                               'src/core/lib/gprpp/time_util.h',
                               'src/core/lib/gprpp/unique_type_name.h',
                               'src/core/lib/gprpp/work_serializer.h',
@@ -2239,7 +2237,6 @@ Pod::Spec.new do |s|
                               'src/core/lib/iomgr/tcp_server.h',
                               'src/core/lib/iomgr/tcp_server_utils_posix.h',
                               'src/core/lib/iomgr/tcp_windows.h',
-                              'src/core/lib/iomgr/time_averaged_stats.h',
                               'src/core/lib/iomgr/timer.h',
                               'src/core/lib/iomgr/timer_generic.h',
                               'src/core/lib/iomgr/timer_heap.h',
