+++ grpc.gemspec
@@ -963,8 +963,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/event_engine/iomgr_engine/iomgr_engine.h )
   s.files += %w( src/core/lib/event_engine/iomgr_engine/thread_pool.cc )
   s.files += %w( src/core/lib/event_engine/iomgr_engine/thread_pool.h )
-  s.files += %w( src/core/lib/event_engine/iomgr_engine/time_averaged_stats.cc )
-  s.files += %w( src/core/lib/event_engine/iomgr_engine/time_averaged_stats.h )
   s.files += %w( src/core/lib/event_engine/iomgr_engine/timer.cc )
   s.files += %w( src/core/lib/event_engine/iomgr_engine/timer.h )
   s.files += %w( src/core/lib/event_engine/iomgr_engine/timer_heap.cc )
@@ -1059,6 +1057,8 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/gprpp/thd_windows.cc )
   s.files += %w( src/core/lib/gprpp/time.cc )
   s.files += %w( src/core/lib/gprpp/time.h )
+  s.files += %w( src/core/lib/gprpp/time_averaged_stats.cc )
+  s.files += %w( src/core/lib/gprpp/time_averaged_stats.h )
   s.files += %w( src/core/lib/gprpp/time_util.cc )
   s.files += %w( src/core/lib/gprpp/time_util.h )
   s.files += %w( src/core/lib/gprpp/unique_type_name.h )
@@ -1189,8 +1189,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/iomgr/tcp_server_windows.cc )
   s.files += %w( src/core/lib/iomgr/tcp_windows.cc )
   s.files += %w( src/core/lib/iomgr/tcp_windows.h )
-  s.files += %w( src/core/lib/iomgr/time_averaged_stats.cc )
-  s.files += %w( src/core/lib/iomgr/time_averaged_stats.h )
   s.files += %w( src/core/lib/iomgr/timer.cc )
   s.files += %w( src/core/lib/iomgr/timer.h )
   s.files += %w( src/core/lib/iomgr/timer_generic.cc )
