+++ grpc.gyp
@@ -802,7 +802,6 @@
         'src/core/lib/event_engine/event_engine.cc',
         'src/core/lib/event_engine/iomgr_engine/iomgr_engine.cc',
         'src/core/lib/event_engine/iomgr_engine/thread_pool.cc',
-        'src/core/lib/event_engine/iomgr_engine/time_averaged_stats.cc',
         'src/core/lib/event_engine/iomgr_engine/timer.cc',
         'src/core/lib/event_engine/iomgr_engine/timer_heap.cc',
         'src/core/lib/event_engine/iomgr_engine/timer_manager.cc',
@@ -813,6 +812,7 @@
         'src/core/lib/event_engine/trace.cc',
         'src/core/lib/gprpp/status_helper.cc',
         'src/core/lib/gprpp/time.cc',
+        'src/core/lib/gprpp/time_averaged_stats.cc',
         'src/core/lib/gprpp/work_serializer.cc',
         'src/core/lib/http/format_request.cc',
         'src/core/lib/http/httpcli.cc',
@@ -880,7 +880,6 @@
         'src/core/lib/iomgr/tcp_server_utils_posix_noifaddrs.cc',
         'src/core/lib/iomgr/tcp_server_windows.cc',
         'src/core/lib/iomgr/tcp_windows.cc',
-        'src/core/lib/iomgr/time_averaged_stats.cc',
         'src/core/lib/iomgr/timer.cc',
         'src/core/lib/iomgr/timer_generic.cc',
         'src/core/lib/iomgr/timer_heap.cc',
@@ -1263,7 +1262,6 @@
         'src/core/lib/event_engine/event_engine.cc',
         'src/core/lib/event_engine/iomgr_engine/iomgr_engine.cc',
         'src/core/lib/event_engine/iomgr_engine/thread_pool.cc',
-        'src/core/lib/event_engine/iomgr_engine/time_averaged_stats.cc',
         'src/core/lib/event_engine/iomgr_engine/timer.cc',
         'src/core/lib/event_engine/iomgr_engine/timer_heap.cc',
         'src/core/lib/event_engine/iomgr_engine/timer_manager.cc',
@@ -1274,6 +1272,7 @@
         'src/core/lib/event_engine/trace.cc',
         'src/core/lib/gprpp/status_helper.cc',
         'src/core/lib/gprpp/time.cc',
+        'src/core/lib/gprpp/time_averaged_stats.cc',
         'src/core/lib/gprpp/work_serializer.cc',
         'src/core/lib/http/format_request.cc',
         'src/core/lib/http/httpcli.cc',
@@ -1340,7 +1339,6 @@
         'src/core/lib/iomgr/tcp_server_utils_posix_noifaddrs.cc',
         'src/core/lib/iomgr/tcp_server_windows.cc',
         'src/core/lib/iomgr/tcp_windows.cc',
-        'src/core/lib/iomgr/time_averaged_stats.cc',
         'src/core/lib/iomgr/timer.cc',
         'src/core/lib/iomgr/timer_generic.cc',
         'src/core/lib/iomgr/timer_heap.cc',
