+++ src/python/grpcio/grpc_core_dependencies.py
@@ -444,7 +444,6 @@ CORE_SOURCE_FILES = [
     'src/core/lib/event_engine/event_engine.cc',
     'src/core/lib/event_engine/iomgr_engine/iomgr_engine.cc',
     'src/core/lib/event_engine/iomgr_engine/thread_pool.cc',
-    'src/core/lib/event_engine/iomgr_engine/time_averaged_stats.cc',
     'src/core/lib/event_engine/iomgr_engine/timer.cc',
     'src/core/lib/event_engine/iomgr_engine/timer_heap.cc',
     'src/core/lib/event_engine/iomgr_engine/timer_manager.cc',
@@ -495,6 +494,7 @@ CORE_SOURCE_FILES = [
     'src/core/lib/gprpp/thd_posix.cc',
     'src/core/lib/gprpp/thd_windows.cc',
     'src/core/lib/gprpp/time.cc',
+    'src/core/lib/gprpp/time_averaged_stats.cc',
     'src/core/lib/gprpp/time_util.cc',
     'src/core/lib/gprpp/work_serializer.cc',
     'src/core/lib/http/format_request.cc',
@@ -563,7 +563,6 @@ CORE_SOURCE_FILES = [
     'src/core/lib/iomgr/tcp_server_utils_posix_noifaddrs.cc',
     'src/core/lib/iomgr/tcp_server_windows.cc',
     'src/core/lib/iomgr/tcp_windows.cc',
-    'src/core/lib/iomgr/time_averaged_stats.cc',
     'src/core/lib/iomgr/timer.cc',
     'src/core/lib/iomgr/timer_generic.cc',
     'src/core/lib/iomgr/timer_heap.cc',
