+++ tools/doxygen/Doxyfile.c++.internal
@@ -1946,8 +1946,6 @@ src/core/lib/event_engine/iomgr_engine/iomgr_engine.cc \
 src/core/lib/event_engine/iomgr_engine/iomgr_engine.h \
 src/core/lib/event_engine/iomgr_engine/thread_pool.cc \
 src/core/lib/event_engine/iomgr_engine/thread_pool.h \
-src/core/lib/event_engine/iomgr_engine/time_averaged_stats.cc \
-src/core/lib/event_engine/iomgr_engine/time_averaged_stats.h \
 src/core/lib/event_engine/iomgr_engine/timer.cc \
 src/core/lib/event_engine/iomgr_engine/timer.h \
 src/core/lib/event_engine/iomgr_engine/timer_heap.cc \
@@ -2042,6 +2040,8 @@ src/core/lib/gprpp/thd_posix.cc \
 src/core/lib/gprpp/thd_windows.cc \
 src/core/lib/gprpp/time.cc \
 src/core/lib/gprpp/time.h \
+src/core/lib/gprpp/time_averaged_stats.cc \
+src/core/lib/gprpp/time_averaged_stats.h \
 src/core/lib/gprpp/time_util.cc \
 src/core/lib/gprpp/time_util.h \
 src/core/lib/gprpp/unique_type_name.h \
@@ -2172,8 +2172,6 @@ src/core/lib/iomgr/tcp_server_utils_posix_noifaddrs.cc \
 src/core/lib/iomgr/tcp_server_windows.cc \
 src/core/lib/iomgr/tcp_windows.cc \
 src/core/lib/iomgr/tcp_windows.h \
-src/core/lib/iomgr/time_averaged_stats.cc \
-src/core/lib/iomgr/time_averaged_stats.h \
 src/core/lib/iomgr/timer.cc \
 src/core/lib/iomgr/timer.h \
 src/core/lib/iomgr/timer_generic.cc \
